Lesson 2.2: Building a view for an existing element
Add TaskListView, which manages the HTML for an entire task list
Move logic for rendering the list out of document.ready

One of the important points here is the split between initialize() and
render(), and figuring out what ought to go where. Remember that
initialize() is called once, when the view is first created, while
render() may be called multiple times. This means that any heavy setup
work (compiling templates, initializing sub-views, finding interesting
sub-elements) should be done in initialize(), and render() should be
kept as skinny as possible.