Lesson 2.3: Organizing views in files
This has nothing to to whatsoever with Backbone, things were starting to
get unwieldy in app.js is all.
Fortunately, Charles' slick project setup allows us to just do this!

This change looks big, but it's mostly just moving code around. The
exception is at the beginning and end of our files, where we've added
export and import commands.

Export makes something defined in this .js file available in another
file. The 'default' keyword means that this is the "main" thing we're
exporting from this file, and makes importing it easier.

Import allows you to access something defined in another .js file.