Lesson 3.1: New Task Event
This one is pretty big. It's also the first time we've added new
functionality to the page since lesson 1.

We've added a form to our HTML, and wired up our TaskListView to listen
for submit events on the form.

If you're not familiar with submit events, they're triggered when the
user clicks on the submit button or presses enter. You can listen for
them in jQuery too, but that's sooooo last week.

Registering an event handler in Backbone is a little different than
jQuery - you use the view's `events` property, matching these weird
event strings to functions (which for some reason are also put down as
strings).

The event handler is pretty similar to what we've seen before: it
consumes the form data, adds it to the task list and re-renders itself.

There are also a couple helper methods in there around interacting with
the form to keep things squeaky clean.