Lesson 3.2: Add a "clear" button
This one is really straightforward - we get it almost for free!

The one trick is that, if you add a button to a form and you don't want
it to trigger submit events, you have to give it
        type="button"

Because this one is simpler, it might make sense to do it first?