Lesson 4.2: Marking tasks complete
Add a field to our task model to track whether it's complete. Also add a
"toggleComplete" function, which reverses the value of complete (true ->
false, false -> true). Just like in Rails, the model is the correct
place for this business logic.

Add some HTML to our template around completion: if a task is complete,
its name should have a strike through it, and there's now a button to
mark a task complete.

Add two event listeners to the TaskView.

The first is a listener on the model, which lands on render. If the
model changes, the view will be notified automatically and re-render
itself.

The second is a listener on the button, so that we can actually toggle
the completion status in the model when the button is pressed.

This is an excellent demonstration of how the Backbone view acts like a
Rails controller. It receives a user event (the click), and does the
right thing (calling toggleComplete in the model). It then receives a
model event, and updates the displayed HTML accordingly.