Lesson 5.2: Deleting Tasks
Add a delete button to the task template.

Add a handler in the TaskView for the new button, which deletes the
model associated with this view. This will remove the model from its
collection, triggering 'remove' and 'update' events in that order.

Add a handler to the TaskListView, bound to the collection's 'remove'
event, which finds the TaskView associated with the removed model and
removes it from the list of views. Because this is guaranteed to happen
before the 'update' event, we don't need to make any changes to
TaskListView.render() - the list will simply be re-rendered without the
removed task.